Fix: alert.SUCCESS() -> alert.success() 수정
+ 로직을 index.js로 변경한 것은 form 등록 후 fetchData로 데이터 재호출하기 위함
